@@ -5,42 +5,27 @@ namespace App\Entity;
 use App\Repository\AlertRepository;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Table(name="alerts",
- *    uniqueConstraints={
- *        @ORM\UniqueConstraint(name="user_name", columns={"user_id", "name"})
- *    }
- * )
- * @ORM\Entity(repositoryClass=AlertRepository::class)
- */
-class Alert
+#[ORM\Table(name: "alerts")]
+#[ORM\UniqueConstraint(name: "user_name", columns: ["user_id", "name"])]
+#[ORM\Entity(repositoryClass: AlertRepository::class)]
+final class Alert
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer", nullable=false)
-     */
-    private string $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: "integer", nullable: false)]
+    private ?string $id = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="User")
-     * @ORM\JoinColumn(name="user_id", referencedColumnName="id")
-     */
+    #[ORM\ManyToOne(targetEntity: User::class)]
+    #[ORM\JoinColumn(name: "user_id", referencedColumnName: "id")]
     private User $user;
 
-    /**
-     * @ORM\Column(type="string", length=255, nullable=false)
-     */
+    #[ORM\Column(name: "name", type: "string", length: 255, nullable: false)]
     private string $name;
 
-    /**
-     * @ORM\Column(name="file", type="string", length=255, nullable=false)
-     */
+    #[ORM\Column(name: "file", type: "string", length: 255, nullable: false)]
     private string $file;
 
-    /**
-     * @ORM\Column(name="active", type="boolean", nullable=false, options={"default":false})
-     */
+    #[ORM\Column(name: "active", type: "boolean", nullable: false, options: ['default' => false])]
     private bool $active = false;
 
     public function __construct(User $user, string $name, string $file = '')
@@ -55,7 +40,7 @@ class Alert
         return $this->id;
     }
 
-    public function getUser(): user
+    public function getUser(): User
     {
         return $this->user;
     }
